# docs/evidence/contracts/iskai-openapi.current.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Centralized AI-Powered Thesis Library using Retrieval-Augmented Generation for the College of Computing Studies, Information and Communication Technology, Isabela State University, Echague.",
    "title": "ISU Thesis AI Library API",
    "version": "2.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/maintenance/alerts": {
      "get": {
        "operationId": "list_operational_alerts_maintenance_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Operational Alerts",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/maintenance/alerts/{alert_id}/acknowledge": {
      "post": {
        "operationId": "acknowledge_alert_maintenance_alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Acknowledge Alert",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/maintenance/operations/summary": {
      "get": {
        "operationId": "operations_summary_maintenance_operations_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Operations Summary",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/maintenance/retention/report": {
      "get": {
        "operationId": "get_retention_report_maintenance_retention_report_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Retention Report",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/maintenance/retention/run": {
      "post": {
        "operationId": "run_retention_maintenance_retention_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "apply",
            "required": false,
            "schema": {
              "default": false,
              "title": "Apply",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The request conflicts with the current state of the resource."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Retention",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/maintenance/storage-cleanup": {
      "get": {
        "operationId": "list_pending_storage_cleanup_maintenance_storage_cleanup_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Pending Storage Cleanup",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/maintenance/storage-cleanup/{task_id}/retry": {
      "post": {
        "operationId": "retry_storage_cleanup_maintenance_storage_cleanup__task_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Storage Cleanup",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/maintenance/upload-jobs": {
      "get": {
        "operationId": "list_upload_jobs_maintenance_upload_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Upload Jobs",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/maintenance/workers": {
      "get": {
        "operationId": "list_workers_maintenance_workers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Workers",
        "tags": [
          "maintenance"
        ]
      }
    },
...
  }
}